new version, allgames screen done

+++ app/src/main/java/com/example/soccergamesfinder/ui/screens/field/FieldDetailsState.kt
@@ -8,7 +8,7 @@ data class FieldDetailsState(
     val games: List<Game> = emptyList(),
     val allGames: List<Game> = emptyList(),
     val totalGames: Int = 0,
-    val avgPlayers: Int = 0,
+    val avgPlayers: Double = 0.0,
     val fullGames: Int = 0,
     val openGames: Int = 0,
     val isLoading: Boolean = false,
