+++ app/src/main/java/com/example/soccergamesfinder/ui/screens/field/FieldStatisticsSection.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.unit.dp
 @Composable
 fun FieldStatisticsSection(
     totalGames: Int,
-    avgPlayers: Int,
+    avgPlayers: Double,
     fullGames: Int,
     openGames: Int
 ) {
